[Feature][Server] Add Catalog Tag

+++ datavines-server/src/main/java/io/datavines/server/api/dto/vo/CatalogTableDetailVO.java
@@ -25,5 +25,5 @@ public class CatalogTableDetailVO extends CatalogEntityBaseDetailVO {
 
     private String comment;
 
-    private int columns;
+    private Long columns;
 }
